+++ chakra-ui-steps/src/components/Step/index.tsx
@@ -18,7 +18,10 @@ export interface StepProps extends HTMLChakraProps<'li'> {
   label?: string | React.ReactNode;
   description?: string;
   icon?: React.ComponentType<any>;
+  state?: 'loading' | 'error';
+  checkIcon?: React.ComponentType<any>;
   isCompletedStep?: boolean;
+  isKeepError?: boolean;
 }
 
 // Props which shouldn't be passed to to the Step component from the user
@@ -37,10 +40,13 @@ export const Step = forwardRef<StepProps, 'li'>(
       children,
       description,
       icon,
+      state,
+      checkIcon,
       index,
       isCompletedStep,
       isCurrentStep,
       isLastStep,
+      isKeepError,
       label,
       ...styleProps
     } = props as FullStepProps;
@@ -50,8 +56,8 @@ export const Step = forwardRef<StepProps, 'li'>(
       isError,
       isLoading,
       isLabelVertical,
-      checkIcon,
       errorIcon,
+      checkIcon: defaultCheckIcon,
       onClickStep,
       clickable,
       setWidths,
@@ -70,18 +76,19 @@ export const Step = forwardRef<StepProps, 'li'>(
       }
     };
 
-    const containerRef = React.createRef<HTMLDivElement>();
-
-    React.useEffect(() => {
-      if (containerRef && containerRef.current && setWidths) {
-        setWidths((prev) => {
-          if (prev.length === stepCount) {
-            return [containerRef.current?.offsetWidth || 0];
-          }
-          return [...prev, containerRef.current?.offsetWidth || 0];
-        });
-      }
-    }, [stepIconContainer.width, stepIconContainer.height]);
+    const containerRef: React.RefCallback<HTMLDivElement> = React.useCallback(
+      (node) => {
+        if (node && setWidths) {
+          setWidths((prev) => {
+            if (prev.length === stepCount) {
+              return [node.offsetWidth || 0];
+            }
+            return [...prev, node.offsetWidth || 0];
+          });
+        }
+      },
+      [stepIconContainer.width, stepIconContainer.height]
+    );
 
     return (
       <>
@@ -112,8 +119,15 @@ export const Step = forwardRef<StepProps, 'li'>(
           >
             <chakra.div
               __css={stepIconContainer}
-              aria-current={isCurrentStep ? 'step' : undefined}
-              data-invalid={dataAttr(isCurrentStep && isError)}
+              aria-current={
+                (hasVisited && isKeepError) || isCurrentStep
+                  ? 'step'
+                  : undefined
+              }
+              data-invalid={dataAttr(
+                ((hasVisited && isKeepError) || isCurrentStep) &&
+                  (isError || state === 'error')
+              )}
               data-highlighted={dataAttr(isCompletedStep)}
               data-clickable={dataAttr(clickable)}
             >
@@ -121,14 +135,15 @@ export const Step = forwardRef<StepProps, 'li'>(
                 <StepIcon
                   {...{
                     index,
-                    isError,
-                    isLoading,
+                    isError: isError || state === 'error',
+                    isLoading: isLoading || state === 'loading',
                     isCurrentStep,
                     isCompletedStep,
+                    isKeepError,
                   }}
                   icon={icon}
-                  checkIcon={checkIcon}
                   errorIcon={errorIcon}
+                  checkIcon={checkIcon ?? defaultCheckIcon}
                 />
               </AnimatePresence>
             </chakra.div>