+++ chakra-ui-steps/src/components/StepIcon/index.tsx
@@ -11,7 +11,8 @@ interface StepIconProps {
   isCurrentStep?: boolean;
   isError?: boolean;
   isLoading?: boolean;
-  icon?: IconType;
+  isKeepError?: boolean;
+  icon?: React.ComponentType<any>;
   index: number;
   checkIcon?: IconType;
   errorIcon?: IconType;
@@ -38,6 +39,7 @@ export const StepIcon = forwardRef<StepIconProps, 'div'>((props, ref) => {
     isCurrentStep,
     isError,
     isLoading,
+    isKeepError,
     icon: CustomIcon,
     index,
     checkIcon: CustomCheckIcon,
@@ -56,7 +58,7 @@ export const StepIcon = forwardRef<StepIconProps, 'div'>((props, ref) => {
     () => (CustomIcon ? CustomIcon : null),
     [CustomIcon]
   );
-
+  
   const ErrorIcon = React.useMemo(
     () => (CustomErrorIcon ? CustomErrorIcon : null),
     [CustomErrorIcon]
@@ -69,6 +71,16 @@ export const StepIcon = forwardRef<StepIconProps, 'div'>((props, ref) => {
 
   return React.useMemo(() => {
     if (isCompletedStep) {
+      if (isError && isKeepError) {
+        return (
+          <AnimatedCloseIcon
+            key="icon"
+            color="white"
+            {...animationConfig}
+            style={icon}
+          />
+        );
+      }
       return (
         <MotionFlex key="check-icon" {...animationConfig}>
           <Check color="white" style={icon} />